requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
bottle==0.13.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clr-loader==0.2.6
colorama==0.4.6
decorator==4.4.2
et-xmlfile==1.1.0
ffmpeg-downloader==0.3.0
filelock==3.16.1
fsspec==2024.10.0
googledriver==0.1.7
h11==0.14.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
keyboard==0.13.5
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
proxy_tools==0.1.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pyperclip==1.9.0

PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-vlc==3.0.21203
pythonnet==3.0.4
pytweening==1.2.0
pytz==2024.2
pywebview==5.3.2
pywin32-ctypes==0.2.3
regex==2024.9.11
requests==2.32.3
selenium==4.25.0
selenium-recaptcha==1.0.7
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.2
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.0
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
